remove insomnia-common
