--- packages/insomnia-common/.eslintignore
@@ -1,2 +1,0 @@
-dist
-coverage
