--- packages/insomnia-common/.eslintrc.js
@@ -1,4 +1,0 @@
-/** @type { import('eslint').Linter.Config } */
-module.exports = {
-  extends: '../../.eslintrc.js',
-};
