--- packages/insomnia-common/jest.config.js
@@ -1,4 +1,0 @@
-/** @type { import('@jest/types').Config.InitialOptions } */
-module.exports = {
-  preset: '../../jest-preset.js',
-};
