--- packages/insomnia-common/package.json
@@ -1,33 +1,0 @@
-{
-  "private": true,
-  "name": "insomnia-common",
-  "version": "3.6.1-beta.3",
-  "homepage": "https://insomnia.rest",
-  "description": "Top-level entities and utilities for Insomnia",
-  "author": "Kong <office@konghq.com>",
-  "license": "Apache-2.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kong/insomnia.git",
-    "directory": "packages/insomnia-config"
-  },
-  "bugs": {
-    "url": "https://github.com/kong/insomnia/issues"
-  },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "bootstrap": "npm run build",
-    "lint": "eslint . --ext .js,.ts,.tsx --cache",
-    "lint:fix": "npm run lint -- --fix",
-    "clean": "tsc --build tsconfig.build.json --clean",
-    "prebuild": "npm run clean",
-    "build": "tsc --build tsconfig.build.json"
-  },
-  "devDependencies": {
-    "ajv": "^8.6.2",
-    "type-fest": "^2.12.0",
-    "typescript": "^4.5.5",
-    "typescript-json-schema": "^0.50.1"
-  }
-}
