--- packages/insomnia-common/src/entities/index.ts
@@ -1,2 +1,0 @@
-export * from './hotkeys';
-export * from './settings';
