--- packages/insomnia-common/tsconfig.build.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "composite": true,
-    "outDir": "dist",
-    "rootDir": "src",
-  },
-  "include": [
-    "src",
-  ],
-  "exclude": [
-    "**/*.test.ts",
-  ],
-}
