--- packages/insomnia-common/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "./tsconfig.build.json",
-  "compilerOptions": {
-    "composite": false,
-    "rootDir": ".",
-  },
-  "include": [
-    ".eslintrc.js",
-    "jest.config.js",
-    "src",
-  ],
-  "exclude": [
-    "dist",
-  ],
-}
