+++ packages/insomnia-config/src/entities.ts
@@ -1,17 +1,15 @@
-import { Settings } from 'insomnia-common';
-
 /** While, by design, there is presently no specific restriction in the validation that _requires_ that the file have a particular filename, this is the canonical name of the file that we use in all current use-cases. */
 export const INSOMNIA_CONFIG_FILENAME = 'insomnia.config.json';
 
 export type ConfigVersion = '1.0.0';
 
-export type AllowedSettings = Partial<Pick<Settings,
-  | 'allowNotificationRequests'
-  | 'disableUpdateNotification'
-  | 'enableAnalytics'
-  | 'disablePaidFeatureAds'
-  | 'incognitoMode'
->>;
+export interface AllowedSettings {
+  allowNotificationRequests: boolean;
+  disableUpdateNotification: boolean;
+  enableAnalytics: boolean;
+  disablePaidFeatureAds: boolean;
+  incognitoMode: boolean;
+}
 
 /**
  * Configuration for Insomnia.
