+++ packages/insomnia-send-request/package.json
@@ -22,7 +22,6 @@
     "hkdf": "0.0.2",
     "html-entities": "^1.3.1",
     "httpsnippet": "^1.22.0",
-    "insomnia-common": "^3.6.1-beta.3",
     "insomnia-config": "^3.6.1-beta.3",
     "insomnia-importers": "^3.6.1-beta.3",
     "insomnia-testing": "^3.6.1-beta.3",
