+++ packages/insomnia/package.json
@@ -64,7 +64,6 @@
     "html-entities": "^1.2.0",
     "httpsnippet": "^2.0.0",
     "iconv-lite": "^0.6.3",
-    "insomnia-common": "^3.6.1-beta.3",
     "insomnia-config": "^3.6.1-beta.3",
     "insomnia-importers": "^3.6.1-beta.3",
     "insomnia-plugin-base64": "^3.6.1-beta.3",
