+++ packages/insomnia/src/common/constants.ts
@@ -1,9 +1,9 @@
-import { KeyCombination } from 'insomnia-common';
 import { unreachableCase } from 'ts-assert-unreachable';
 
 import appConfig from '../../config/config.json';
 import { version } from '../../package.json';
 import { getPortableExecutableDir } from './electron-helpers';
+import { KeyCombination } from './settings';
 
 const env = process['env'];
