+++ packages/insomnia/src/common/settings.ts
@@ -1,5 +1,81 @@
-import { HttpVersion } from '../constants';
-import { HotKeyRegistry } from './hotkeys';
+import { ValueOf } from 'type-fest';
+
+/**
+ * The readable definition of a hotkey.
+ */
+export interface KeyboardShortcutDefinition {
+  id: string;
+  description: string;
+}
+
+/**
+ * The combination of key presses that will activate a hotkey if pressed.
+ */
+export interface KeyCombination {
+  ctrl?: boolean;
+  alt?: boolean;
+  shift?: boolean;
+  meta?: boolean;
+  keyCode: number;
+}
+
+/**
+ * The collection of a hotkey's key combinations for each platforms.
+ */
+export interface PlatformKeyCombinations {
+  macKeys: KeyCombination[];
+  // The key combinations for both Windows and Linux.
+  winLinuxKeys: KeyCombination[];
+}
+
+export type KeyboardShortcut =
+  | 'workspace_showSettings'
+  | 'request_showSettings'
+  | 'preferences_showKeyboardShortcuts'
+  | 'preferences_showGeneral'
+  | 'request_quickSwitch'
+  | 'request_showRecent'
+  | 'request_showRecentPrevious'
+  | 'plugin_reload'
+  | 'showAutocomplete'
+  | 'request_send'
+  | 'request_showOptions'
+  | 'environment_showEditor'
+  | 'environment_showSwitchMenu'
+  | 'request_toggleHttpMethodMenu'
+  | 'request_toggleHistory'
+  | 'request_focusUrl'
+  | 'request_showGenerateCodeEditor'
+  | 'sidebar_focusFilter'
+  | 'sidebar_toggle'
+  | 'response_focus'
+  | 'showCookiesEditor'
+  | 'request_createHTTP'
+  | 'request_showDelete'
+  | 'request_showCreateFolder'
+  | 'request_showDuplicate'
+  | 'request_togglePin'
+  | 'environment_showVariableSourceAndValue'
+  | 'beautifyRequestBody'
+  | 'graphql_explorer_focus_filter';
+
+/**
+ * The collection of defined hotkeys.
+ * The registry maps a hotkey by its reference id to its key bindings.
+ */
+export type HotKeyRegistry = Record<KeyboardShortcut, PlatformKeyCombinations>;
+
+// HTTP version codes
+export const HttpVersions = {
+  V1_0: 'V1_0',
+  V1_1: 'V1_1',
+  V2PriorKnowledge: 'V2PriorKnowledge',
+  V2_0: 'V2_0',
+  v3: 'v3',
+  default: 'default',
+} as const;
+
+export type HttpVersion = ValueOf<typeof HttpVersions>;
 
 export enum UpdateChannel {
   stable = 'stable',
