+++ packages/insomnia/src/models/helpers/settings.ts
@@ -1,11 +1,11 @@
 import { readFileSync } from 'fs';
-import { Settings } from 'insomnia-common';
 import { ErrorResult, INSOMNIA_CONFIG_FILENAME, InsomniaConfig, isErrorResult, validate } from 'insomnia-config';
 import { resolve } from 'path';
 import type { ValueOf } from 'type-fest';
 
 import { isDevelopment } from '../../common/constants';
 import { getDataDirectory, getPortableExecutableDir } from '../../common/electron-helpers';
+import { Settings } from '../../common/settings';
 
 interface FailedParseResult {
   syntaxError: SyntaxError;
