+++ packages/insomnia/src/ui/components/panes/placeholder-response-pane.tsx
@@ -1,4 +1,4 @@
-import { KeyboardShortcut } from 'insomnia-common';
+import { KeyboardShortcut } from '../../../common/settings';
 import React, { FC, PropsWithChildren } from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
