+++ packages/insomnia/src/ui/components/settings/masked-setting.tsx
@@ -1,8 +1,8 @@
-import { SettingsOfType } from 'insomnia-common';
 import React, { ChangeEvent, FC } from 'react';
 import { useSelector } from 'react-redux';
 import { useToggle } from 'react-use';
 
+import { SettingsOfType } from '../../../common/settings';
 import * as models from '../../../models';
 import { selectSettings } from '../../redux/selectors';
 import { HelpTooltip } from '../help-tooltip';
