+++ packages/insomnia/src/ui/components/settings/number-setting.tsx
@@ -1,8 +1,8 @@
-import { SettingsOfType } from 'insomnia-common';
 import React, { ChangeEventHandler, FC, InputHTMLAttributes, useCallback } from 'react';
 import { useSelector } from 'react-redux';
 
 import { snapNumberToLimits } from '../../../common/misc';
+import { SettingsOfType } from '../../../common/settings';
 import * as models from '../../../models/index';
 import { selectSettings } from '../../redux/selectors';
 import { HelpTooltip } from '../help-tooltip';
