+++ packages/insomnia/src/ui/components/settings/text-setting.tsx
@@ -1,7 +1,7 @@
-import { SettingsOfType } from 'insomnia-common';
 import React, { ChangeEventHandler, FC, InputHTMLAttributes, useCallback } from 'react';
 import { useSelector } from 'react-redux';
 
+import { SettingsOfType } from '../../../common/settings';
 import * as models from '../../../models/index';
 import { selectSettings } from '../../redux/selectors';
 import { HelpTooltip } from '../help-tooltip';
